✨feat: separated ReviewMorePageTemplate
Related to: #117
